Shooting Arrows
Can shoot arrows that have a rigidbody on them (need to tweak). Currently countering race condition by resetting a bool within certain animation events in order to properly destroy the arrow if it is being shot or not. Wrapped the spine rotation method and logic within RotatePlayer to allow Root motion to be applied while the player is dodging.